chore: Rename FoxRiveLoaderAnimation mock to play well with local jest unit tests
