+++ app/components/Views/ChoosePassword/FoxRiveLoaderAnimation/__mocks__/FoxRiveLoaderAnimation.tsx
